+++ src/basic/DAXPY_ATOMIC.cpp
@@ -61,7 +61,7 @@ DAXPY_ATOMIC::~DAXPY_ATOMIC()
 
 void DAXPY_ATOMIC::setUp(VariantID vid, size_t RAJAPERF_UNUSED_ARG(tune_idx))
 {
-  allocAndInitDataConst(m_y, getActualProblemSize(), 0.0, vid);
+  allocAndInitDataConst(m_y, getActualProblemSize(), (float) 0.0, vid);
   allocAndInitData(m_x, getActualProblemSize(), vid);
   initData(m_a, vid);
 }
