+++ src/basic/INIT_VIEW1D.cpp
@@ -62,7 +62,7 @@ INIT_VIEW1D::~INIT_VIEW1D()
 
 void INIT_VIEW1D::setUp(VariantID vid, size_t RAJAPERF_UNUSED_ARG(tune_idx))
 {
-  allocAndInitDataConst(m_a, getActualProblemSize(), 0.0, vid);
+  allocAndInitDataConst(m_a, getActualProblemSize(), (float) 0.0, vid);
   m_val = 0.00000123;
 }
