+++ src/lcals/EOS.cpp
@@ -67,7 +67,7 @@ EOS::~EOS()
 
 void EOS::setUp(VariantID vid, size_t RAJAPERF_UNUSED_ARG(tune_idx))
 {
-  allocAndInitDataConst(m_x, m_array_length, 0.0, vid);
+  allocAndInitDataConst(m_x, m_array_length, (float) 0.0, vid);
   allocAndInitData(m_y, m_array_length, vid);
   allocAndInitData(m_z, m_array_length, vid);
   allocAndInitData(m_u, m_array_length, vid);
