+++ src/polybench/POLYBENCH_ATAX.cpp
@@ -77,7 +77,7 @@ void POLYBENCH_ATAX::setUp(VariantID vid, size_t RAJAPERF_UNUSED_ARG(tune_idx))
   allocAndInitData(m_tmp, m_N, vid);
   allocAndInitData(m_x, m_N, vid);
   allocAndInitData(m_A, m_N * m_N, vid);
-  allocAndInitDataConst(m_y, m_N, 0.0, vid);
+  allocAndInitDataConst(m_y, m_N, (float) 0.0, vid);
 }
 
 void POLYBENCH_ATAX::updateChecksum(VariantID vid, size_t tune_idx)
