+++ src/polybench/POLYBENCH_GEMM.cpp
@@ -81,7 +81,7 @@ void POLYBENCH_GEMM::setUp(VariantID vid, size_t RAJAPERF_UNUSED_ARG(tune_idx))
   (void) vid;
   allocAndInitData(m_A, m_ni * m_nk, vid);
   allocAndInitData(m_B, m_nk * m_nj, vid);
-  allocAndInitDataConst(m_C, m_ni * m_nj, 0.0, vid);
+  allocAndInitDataConst(m_C, m_ni * m_nj, (float) 0.0, vid);
 }
 
 void POLYBENCH_GEMM::updateChecksum(VariantID vid, size_t tune_idx)
