+++ src/stream/MUL.cpp
@@ -62,7 +62,7 @@ MUL::~MUL()
 
 void MUL::setUp(VariantID vid, size_t RAJAPERF_UNUSED_ARG(tune_idx))
 {
-  allocAndInitDataConst(m_b, getActualProblemSize(), 0.0, vid);
+  allocAndInitDataConst(m_b, getActualProblemSize(), (float) 0.0, vid);
   allocAndInitData(m_c, getActualProblemSize(), vid);
   initData(m_alpha, vid);
 }
